timeMsg double
git-svn-id: svn://elaine/murooma/trunk@266 d8a302eb-03bc-478d-80e4-98257eca68ef

+++ common/timeMsg.h
@@ -22,7 +22,7 @@ public:
 
 	virtual uint32_t getSize()
 	{
-		return sizeof(int32_t);
+		return sizeof(double);
 	}
 
 	double latency;
